# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Corredores",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/corredor/{id}": {
            "put": {
                "tags": [
                    "Corredor"
                ],
                "summary": "Atualiza corredor",
                "operationId": "App\\Http\\Controllers\\CorredorController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID corredor",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Exemplo para atualizar um corredor",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "cpf",
                                    "data_nascimento"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Nome do candidato"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "00000000000"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1994-11-07"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Operação realizada com sucesso."
                                        },
                                        "sucesso": {
                                            "type": "bool",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Exemplos de possíveis erros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "object",
                                            "example": {
                                                "nome": [
                                                    "O campo nome não pode ser superior a 255 caracteres.",
                                                    "O campo nome deve ser uma string."
                                                ],
                                                "cpf": [
                                                    "O campo cpf já está sendo utilizado.",
                                                    "O campo cpf não pode ser superior a 11 caracteres.",
                                                    "O campo cpf deve ter pelo menos 11 caracteres."
                                                ],
                                                "data_nascimento": [
                                                    "Inscrição permitida apenas para maiores de 18 anos.",
                                                    "O campo data nascimento não é uma data válida."
                                                ]
                                            }
                                        },
                                        "sucesso": {
                                            "type": "bool",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Corredor"
                ],
                "summary": "Deleta corredor",
                "operationId": "App\\Http\\Controllers\\CorredorController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID corredor",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Operação realizada com sucesso."
                                        },
                                        "sucesso": {
                                            "type": "bool",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Exemplos de possíveis erros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Não é possível excluir o corredor, pois o mesmo possui cadastro para uma ou mais provas"
                                        },
                                        "sucesso": {
                                            "type": "bool",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/corredor-em-prova/{id}": {
            "delete": {
                "tags": [
                    "Corredor em Prova"
                ],
                "summary": "Deleta um corredor de uma prova",
                "operationId": "App\\Http\\Controllers\\CorredorEmProvaController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID corredor em prova",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Operação realizada com sucesso."
                                        },
                                        "sucesso": {
                                            "type": "bool",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Exemplos de possíveis erros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Não é possível excluir o corredor desta prova, pois o mesmo já possui resultado gerado."
                                        },
                                        "sucesso": {
                                            "type": "bool",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/prova/{id}": {
            "put": {
                "tags": [
                    "Prova"
                ],
                "summary": "Atualiza prova",
                "operationId": "App\\Http\\Controllers\\ProvaController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID prova",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Exemplo para atualizar uma prova",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tipo_prova_id",
                                    "data"
                                ],
                                "properties": {
                                    "tipo_prova_id": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    },
                                    "data": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2021-05-31 10:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Operação realizada com sucesso."
                                        },
                                        "sucesso": {
                                            "type": "bool",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Exemplos de possíveis erros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "object",
                                            "example": {
                                                "tipo_prova_id": [
                                                    "O campo tipo prova id selecionado é inválido."
                                                ],
                                                "data_nascimento": [
                                                    "O campo data não é uma data válida."
                                                ]
                                            }
                                        },
                                        "sucesso": {
                                            "type": "bool",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Prova"
                ],
                "summary": "Deleta prova",
                "operationId": "App\\Http\\Controllers\\ProvaController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID prova",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Operação realizada com sucesso."
                                        },
                                        "sucesso": {
                                            "type": "bool",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Exemplos de possíveis erros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Não é possível excluir a prova, pois existem corredores cadastrados para ela."
                                        },
                                        "sucesso": {
                                            "type": "bool",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}